implement CellValueStream

+++ src/main/kotlin/DSCUSBDrv64.kt
@@ -10,7 +10,7 @@ interface DSCUSBDrv64 {
     fun WRITECOMMAND(@In Serial: Int, @In Command: String, @In Value: Float): Int
     fun EXECUTECOMMAND(@In Serial: Int, @In Command: String): Int
     fun SETTIMEOUT(@In NewTimeout: Int): Int
-    fun GETTIMEOUT(): Int
+    fun GETTIMEOUT(): Long
     fun VERSION(): Float
     fun DEVICECOUNT(): Int
     fun SERIALNUMBER(@In Index: Byte): Int
